Hard code generated Rust sources into the library.
